Copied SD card driver into the project. It compiles at least and *says* that it initialized the card correctly.
The driver depends on a libary called "tock-registers" which had some dependency issues with the version of nightly rust that we were using, so I simply copied it in and modified it to work. This should be seen as a stop-gap solution until we upgrade to a new version of stable rust.